Project2_Discord_Bot/requirements.txt
# pip requirements file

# Discord API client
discord.py

# dotenv to parse environment var file
python-dotenv

# flashgeotext for city names
flashgeotext

# to make API calls
requests
